Get municipality congruences

Returns a tibble with congruence data of Swiss municipalities.

congruences(
  start_date = NULL,
  end_date = NULL,
  incl_unmodified = TRUE,
  incl_territory_exchange = FALSE,
  use_cache = TRUE,
  cache_lifespan = "1 day"
)

Arguments

start_date
The begin of the period the data covers. A date or a character scalar in the format "YYYY-MM-DD".

end_date
The end of the period the data covers. A date or a character scalar in the format "YYYY-MM-DD".

incl_unmodified
Include municipalities where no changes occurred in the defined period.

incl_territory_exchange
Include information (additional rows) about territory exchanges within and between municipalities.

use_cache
Return cached results if possible. If FALSE, results are always newly fetched regardless of cache_lifespan.

cache_lifespan
The duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Only relevant if use_cache = TRUE. Defaults to 1 day (24 hours).

Value

A tibble.

Examples

swissmuni::congruences(start_date = lubridate::today(),
                       end_date = lubridate::today(),
                       cache_lifespan = "6 hours")
#> # A tibble: 2,148 × 12
#>    InitialHistoricalC… InitialCode InitialName InitialParentHi… InitialParentNa…
#>                  <int>       <int> <chr>                  <int> <chr>           
#>  1               10009        4122 Villnachern            10023 Bezirk Brugg    
#>  2               10078        6158 Vionnaz                10013 District de Mon…
#>  3               10157        3023 Speicher               10098 Bezirk Mittella…
#>  4               10159        6011 Zwischberg…            10035 Bezirk Brig     
#>  5               10162        2228 Villars-su…            10104 District de la …
#>  6               10165        2230 Villarsel-…            10104 District de la …
#>  7               10242         181 Wila                   10076 Bezirk Pfäffikon
#>  8               10268          71 Wil (ZH)               10081 Bezirk Bülach   
#>  9               10275        4045 Wettingen              10025 Bezirk Baden    
#> 10               10277         121 Wetzikon (…            10079 Bezirk Hinwil   
#> # … with 2,138 more rows, and 7 more variables: InitialStep <int>,
#> #   TerminalHistoricalCode <int>, TerminalCode <int>, TerminalName <chr>,
#> #   TerminalParentHistoricalCode <int>, TerminalParentName <chr>,
#> #   TerminalStep <int>