Get spatial classifications of municipalities

Returns a tibble with a wide range of spatial classifications of Swiss municipalities, i.a. by language region, size of the municipality or degree of agglomeration/urbanization.

Usage

classifications(
  start_date = NULL,
  end_date = NULL,
  historicized_code = FALSE,
  name_type = c("ID", "en", "de", "fr", "it"),
  use_cache = TRUE,
  cache_lifespan = "1 day"
)

Arguments

start_date

The begin of the period the data covers. A date or a character scalar in the format "YYYY-MM-DD".

end_date

The end of the period the data covers. A date or a character scalar in the format "YYYY-MM-DD".

historicized_code

By default, the FSO commune number is returned. Set to TRUE in order to get the historicization number instead.

name_type

The column name type to be returned. Defaults to the column IDs returned by the web service. If set to a language code, the columns starting with ID HR_ will be named by its full label in the respective language instead.

use_cache

Return cached results if possible. If FALSE, results are always newly fetched regardless of cache_lifespan.

cache_lifespan

The duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Only relevant if use_cache = TRUE. Defaults to 1 day (24 hours).

Value

A tibble.

Details

There are some classifications which have multiple column IDs (e.g. HR_ARBREG2000_L2 and HR_TYPMSR2000_L2 both contain the spatial mobility regions). If name_type is set to a language code, column duplicates are given the name suffix " (⎘)".

FSO commune number vs. historicization number

By default, the FSO commune number is returned. You can set historicized_code = TRUE to get the historicization number instead. Details about the differences between the both can be found in this PDF (in German). A translation of the relevant paragraph (p. 2):

―

The mutation processes at the cantonal, district and municipal level cause changes in the official register of Swiss municipalities. The so-called historicization number uniquely identifies the entries in the official commune register that have been changed, “deleted” or added as a result of mutation processes.

In contrast to the FSO commune number or district number, the historicization number is unique within the database and defines the state of the communes or districts during a certain period of time. The mutation numbers, the type of inclusion or revocation and the date of inclusion or revocation are used to fully describe the individual mutation events in the “Municipalities” and “Districts” tables. The mutation number identifies the individual mutation events that may affect one or more entries.

The historicization number does not replace the existing municipality and district numbers. It is used for the computer-aided mapping of the various states and mutation processes of the official commune register and the districts.

―

Examples

swissmuni::classifications(start_date = lubridate::today(),
                           end_date = lubridate::today(),
                           name_type = "de",
                           use_cache = FALSE)
#> # A tibble: 2,145 × 23
#>    Ident…¹ Name  CODE_…² Kanto…³ Histo…⁴ Agglo…⁵ Agglo…⁶ Agglo…⁷ Arbei…⁸ Arbei…⁹ Urban…˟ Funkt…˟ Erwei…˟ Städt…˟ Gemei…˟ Gemei…˟ Bergg…˟ Gross…˟ Raum …˟ Sprac…˟
#>      <int> <chr>   <int>   <int>   <int>   <int>   <int>   <int>   <int>   <int>   <int>   <int>   <int>   <int>   <int>   <int>   <int>   <int>   <int>   <int>
#>  1   10009 Vill…    4122      19   10023       0       0       0      12   12010       2       0       0       3      23     236       0       3       5       1
#>  2   10078 Vion…    6158      23   10013       0       0       0       2    2040       3       0       0       2      22     226       1       1       5       2
#>  3   10157 Spei…    3023      15   10098    3203       3    3203      15   15050       2 3203960       0       2      22     227       1       5       4       1
#>  4   10159 Zwis…    6011      23   10035    6002       5    6002       7    7090       3       0       0       3      23     237       1       1       4       1
#>  5   10162 Vill…    2228      10   10104    2196       3    2196       4    4030       2       0       0       1      12     122       0       2       2       2
#>  6   10165 Vill…    2230      10   10104    2196       3    2196       4    4030       3       0       0       3      23     235       0       2       4       2
#>  7   10242 Wila      181       1   10076       0       0       0      12   12040       3  261960       0       2      22     226       0       4       5       1
#>  8   10268 Wil …      71       1   10081     261       1     261      12   12032       3  261960       0       3      23     236       0       4       4       1
#>  9   10275 Wett…    4045      19   10025    4021       3    4021      12   12010       2  261960       0       1      12     123       0       3       2       1
#> 10   10277 Wetz…     121       1   10079     261       1     261      12   12037       2  261960       0       1      11     112       0       4       3       1
#> # … with 2,135 more rows, 3 more variables: `Statistische Städte 2012` <int>, INSCRIPTION <int>, ABBREV <chr>, and abbreviated variable names ¹​Identifier,
#> #   ²​CODE_OFS, ³​`Kanton (2022-01-01)`, ⁴​`Historisiertes Bezirkverzeichnis der Schweiz (2022-01-01)`, ⁵​`Agglomerationen 2012`,
#> #   ⁶​`Agglomerationsgrössenklasse 2012`, ⁷​`Agglomerationen 2012 ⎘`, ⁸​`Arbeitsmarktgrossregionen 2018`, ⁹​`Arbeitsmarktregionen 2018`,
#> #   ˟​`Urbanisierungsgrad 2011 (DEGURBA) - eurostat 2011`, ˟​`Funktionale städtische Gebiete 2014 (FUA eurostat)`,
#> #   ˟​`Erweiterte Städte 2011 (Greater cities eurostat)`, ˟​`Städtische / Ländliche Gebiete`, ˟​`Gemeindetypologie 2012 (9 Typen)`,
#> #   ˟​`Gemeindetypologie 2012 (25 Typen)`, ˟​Berggebiete, ˟​`Grossregionen der Schweiz`, ˟​`Raum mit städtischem Charakter 2012`, ˟​`Sprachgebiete 2016`