Get municipality congruences

Returns a tibble with congruence data of Swiss municipalities. The data indicates for each existing municipality at the start_date to which municipality/municipalities it corresponds at the end_date.

Usage

congruences(
  start_date = NULL,
  end_date = NULL,
  incl_unmodified = TRUE,
  incl_territory_exchange = FALSE,
  use_cache = TRUE,
  cache_lifespan = "1 day"
)

Arguments

start_date

The begin of the period the data covers. A date or a character scalar in the format "YYYY-MM-DD".

end_date

The end of the period the data covers. A date or a character scalar in the format "YYYY-MM-DD".

incl_unmodified

Include municipalities where no changes occurred in the defined period.

incl_territory_exchange

Include information (additional rows) about territory exchanges within and between municipalities.

use_cache

Return cached results if possible. If FALSE, results are always newly fetched regardless of cache_lifespan.

cache_lifespan

The duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Only relevant if use_cache = TRUE. Defaults to 1 day (24 hours).

Value

A tibble.

Examples

swissmuni::congruences(start_date = lubridate::today(),
                       end_date = lubridate::today(),
                       cache_lifespan = "6 hours")
#> # A tibble: 2,145 × 12
#>    InitialHistoricalCode InitialCode InitialName         InitialParentHistoricalCode InitialParentName   Initi…¹ Termi…² Termi…³ Termi…⁴ Termi…⁵ Termi…⁶ Termi…⁷
#>                    <int>       <int> <chr>                                     <int> <chr>                 <int>   <int>   <int> <chr>     <int> <chr>     <int>
#>  1                 10009        4122 Villnachern                               10023 Bezirk Brugg             NA   10009    4122 Villna…   10023 Bezirk…      NA
#>  2                 10078        6158 Vionnaz                                   10013 District de Monthey      NA   10078    6158 Vionnaz   10013 Distri…      NA
#>  3                 10157        3023 Speicher                                  10098 Bezirk Mittelland        NA   10157    3023 Speich…   10098 Bezirk…      NA
#>  4                 10159        6011 Zwischbergen                              10035 Bezirk Brig              NA   10159    6011 Zwisch…   10035 Bezirk…      NA
#>  5                 10162        2228 Villars-sur-Glâne                         10104 District de la Sar…      NA   10162    2228 Villar…   10104 Distri…      NA
#>  6                 10165        2230 Villarsel-sur-Marly                       10104 District de la Sar…      NA   10165    2230 Villar…   10104 Distri…      NA
#>  7                 10242         181 Wila                                      10076 Bezirk Pfäffikon         NA   10242     181 Wila      10076 Bezirk…      NA
#>  8                 10268          71 Wil (ZH)                                  10081 Bezirk Bülach            NA   10268      71 Wil (Z…   10081 Bezirk…      NA
#>  9                 10275        4045 Wettingen                                 10025 Bezirk Baden             NA   10275    4045 Wettin…   10025 Bezirk…      NA
#> 10                 10277         121 Wetzikon (ZH)                             10079 Bezirk Hinwil            NA   10277     121 Wetzik…   10079 Bezirk…      NA
#> # … with 2,135 more rows, and abbreviated variable names ¹​InitialStep, ²​TerminalHistoricalCode, ³​TerminalCode, ⁴​TerminalName, ⁵​TerminalParentHistoricalCode,
#> #   ⁶​TerminalParentName, ⁷​TerminalStep