Get municipality mutations

Returns a tibble with mutation data of Swiss municipalities.

mutations(
  start_date = NULL,
  end_date = NULL,
  incl_territory_exchange = FALSE,
  use_cache = TRUE,
  cache_lifespan = "1 day"
)

Arguments

start_date
The begin of the period the data covers. A date or a character scalar in the format "YYYY-MM-DD".

end_date
The end of the period the data covers. A date or a character scalar in the format "YYYY-MM-DD".

incl_territory_exchange
Include information (additional rows) about territory exchanges within and between municipalities.

use_cache
Return cached results if possible. If FALSE, results are always newly fetched regardless of cache_lifespan.

cache_lifespan
The duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Only relevant if use_cache = TRUE. Defaults to 1 day (24 hours).

Value

A tibble.

Examples

swissmuni::mutations(start_date = "2020-01-01",
                     end_date = "2020-06-30",
                     incl_territory_exchange = TRUE)
#> # A tibble: 18 × 14
#>    MutationNumber MutationDate InitialHistoricalCode InitialCode InitialName    
#>             <int> <date>                       <int>       <int> <chr>          
#>  1           3920 2020-01-01                   15297         873 Kirchenthurnen 
#>  2           3920 2020-01-01                   15298         874 Lohnstorf      
#>  3           3920 2020-01-01                   15300         876 Mühlethurnen   
#>  4           3921 2020-01-01                   15334         937 Schwendibach   
#>  5           3921 2020-01-01                   15336         939 Steffisburg    
#>  6           3922 2020-01-01                   15366         981 Niederbipp     
#>  7           3922 2020-01-01                   15378         996 Wolfisberg     
#>  8           3923 2020-01-01                   15541        1123 Altishofen     
#>  9           3923 2020-01-01                   15553        1126 Ebersecken     
#> 10           3924 2020-01-01                   13268        2111 Villaz-Saint-P…
#> 11           3924 2020-01-01                   14513        2116 La Folliaz     
#> 12           3925 2020-01-01                   11633        2185 Corserey       
#> 13           3925 2020-01-01                   12550        2213 Noréaz         
#> 14           3925 2020-01-01                   13103        2221 Prez-vers-Noré…
#> 15           3926 2020-01-01                   16027        3901 Chur           
#> 16           3926 2020-01-01                   16028        3926 Maladers       
#> 17           3927 2020-01-01                   14376        4114 Schinznach-Bad 
#> 18           3927 2020-01-01                   15387        4095 Brugg          
#> # … with 9 more variables: InitialParentHistoricalCode <int>,
#> #   InitialParentName <chr>, InitialStep <int>, TerminalHistoricalCode <int>,
#> #   TerminalCode <int>, TerminalName <chr>, TerminalParentHistoricalCode <int>,
#> #   TerminalParentName <chr>, TerminalStep <int>