Get municipality mutations

Returns a tibble with mutation data of Swiss municipalities. The data describes all changes related to the municipalities made between start_date and end_date.

Usage

mutations(
  start_date = NULL,
  end_date = NULL,
  incl_territory_exchange = FALSE,
  use_cache = TRUE,
  cache_lifespan = "1 day"
)

Arguments

start_date

The begin of the period the data covers. A date or a character scalar in the format "YYYY-MM-DD".

end_date

The end of the period the data covers. A date or a character scalar in the format "YYYY-MM-DD".

incl_territory_exchange

Include information (additional rows) about territory exchanges within and between municipalities.

use_cache

Return cached results if possible. If FALSE, results are always newly fetched regardless of cache_lifespan.

cache_lifespan

The duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Only relevant if use_cache = TRUE. Defaults to 1 day (24 hours).

Value

A tibble.

Examples

swissmuni::mutations(start_date = "2020-01-01",
                     end_date = "2020-06-30",
                     incl_territory_exchange = TRUE)
#> # A tibble: 18 × 14
#>    MutationNumber MutationDate InitialHistoricalCode InitialCode InitialName         InitialPa…¹ Initi…² Initi…³ Termi…⁴ Termi…⁵ Termi…⁶ Termi…⁷ Termi…⁸ Termi…⁹
#>             <int> <date>                       <int>       <int> <chr>                     <int> <chr>     <int>   <int>   <int> <chr>     <int> <chr>     <int>
#>  1           3920 2020-01-01                   15297         873 Kirchenthurnen            10288 Verwal…      29   16128     889 Thurnen   10288 Verwal…      21
#>  2           3920 2020-01-01                   15298         874 Lohnstorf                 10288 Verwal…      29   16128     889 Thurnen   10288 Verwal…      21
#>  3           3920 2020-01-01                   15300         876 Mühlethurnen              10288 Verwal…      29   16128     889 Thurnen   10288 Verwal…      21
#>  4           3921 2020-01-01                   15334         937 Schwendibach              10289 Verwal…      29   16129     939 Steffi…   10289 Verwal…      26
#>  5           3921 2020-01-01                   15336         939 Steffisburg               10289 Verwal…      26   16129     939 Steffi…   10289 Verwal…      26
#>  6           3922 2020-01-01                   15366         981 Niederbipp                10286 Verwal…      26   16130     981 Nieder…   10286 Verwal…      26
#>  7           3922 2020-01-01                   15378         996 Wolfisberg                10286 Verwal…      29   16130     981 Nieder…   10286 Verwal…      26
#>  8           3923 2020-01-01                   15541        1123 Altishofen                10302 Wahlkr…      26   16131    1123 Altish…   10302 Wahlkr…      26
#>  9           3923 2020-01-01                   15553        1126 Ebersecken                10302 Wahlkr…      29   16131    1123 Altish…   10302 Wahlkr…      26
#> 10           3924 2020-01-01                   13268        2111 Villaz-Saint-Pierre       10106 Distri…      29   16132    2117 Villaz    10106 Distri…      21
#> 11           3924 2020-01-01                   14513        2116 La Folliaz                10106 Distri…      29   16132    2117 Villaz    10106 Distri…      21
#> 12           3925 2020-01-01                   11633        2185 Corserey                  10104 Distri…      29   16133    2237 Prez      10104 Distri…      21
#> 13           3925 2020-01-01                   12550        2213 Noréaz                    10104 Distri…      29   16133    2237 Prez      10104 Distri…      21
#> 14           3925 2020-01-01                   13103        2221 Prez-vers-Noréaz          10104 Distri…      29   16133    2237 Prez      10104 Distri…      21
#> 15           3926 2020-01-01                   16027        3901 Chur                      10314 Region…      26   16134    3901 Chur      10314 Region…      26
#> 16           3926 2020-01-01                   16028        3926 Maladers                  10314 Region…      29   16134    3901 Chur      10314 Region…      26
#> 17           3927 2020-01-01                   14376        4114 Schinznach-Bad            10023 Bezirk…      29   16135    4095 Brugg     10023 Bezirk…      26
#> 18           3927 2020-01-01                   15387        4095 Brugg                     10023 Bezirk…      26   16135    4095 Brugg     10023 Bezirk…      26
#> # … with abbreviated variable names ¹​InitialParentHistoricalCode, ²​InitialParentName, ³​InitialStep, ⁴​TerminalHistoricalCode, ⁵​TerminalCode, ⁶​TerminalName,
#> #   ⁷​TerminalParentHistoricalCode, ⁸​TerminalParentName, ⁹​TerminalStep